_experimenting/prototype/requirements.txt
kafka-python
flask
arrow
mypy-lang
Sphinx
sklearn
numpy
pandas
nose